Change createEntityItem to float p.1

+++ CraftTweaker2-MC1120-Main/src/main/java/crafttweaker/mc1120/item/MCItemStack.java
@@ -562,7 +562,7 @@ public class MCItemStack implements IItemStack {
     }
     
     @Override
-    public IEntityItem createEntityItem(IWorld world, int x, int y, int z) {
+    public IEntityItem createEntityItem(IWorld world, float x, float y, float z) {
         return CraftTweakerMC.getIEntityItem(new EntityItem(CraftTweakerMC.getWorld(world), x, y, z, stack));
     }
